Update lecture inside, outside
