+++ packages/outside/src/styles/lecture/style.ts
@@ -1,0 +1,67 @@
+import styled from '@emotion/styled'
+
+export const SlideBg = styled.div<{ url: any }>`
+  height: 15rem;
+  width: 100%;
+  display: flex;
+  justify-content: center;
+  background-image: url(${({ url }) => url.src});
+  background-position: center;
+  background-size: cover;
+  align-items: flex-end;
+`
+
+export const BgContainer = styled.div`
+  width: 75rem;
+  display: flex;
+  justify-content: space-between;
+  margin-bottom: 1rem;
+`
+
+export const LectureTitle = styled.span`
+  color: ${({ theme }) => theme.color.white};
+  ${({ theme }) => theme.typo.title_lg};
+`
+
+export const ButtonContainer = styled.div`
+  display: flex;
+  align-items: flex-end;
+`
+
+export const LectureButton = styled.div`
+  display: flex;
+  align-items: center;
+  cursor: pointer;
+  background-color: rgb(255, 255, 255, 0.2);
+  height: 2.5rem;
+  margin-left: 1rem;
+  padding: 0 0.75rem;
+  border-radius: 0.5rem;
+  backdrop-filter: blur(0.25rem);
+  svg {
+    fill: ${({ theme }) => theme.color.white};
+  }
+  span {
+    color: ${({ theme }) => theme.color.white};
+    ${({ theme }) => theme.typo.text_md};
+    margin-left: 0.25rem;
+  }
+  &:hover {
+    background-color: rgb(255, 255, 255, 0.4);
+  }
+`
+
+export const SelectFilterContainer = styled.div`
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+`
+
+export const ListWrraper = styled.div`
+  display: flex;
+  justify-content: center;
+`
+export const ListContainer = styled.div`
+  width: 75rem;
+  margin-top: 1.5rem;
+`
