+++ shared/common/src/components/Header/index.tsx
@@ -5,7 +5,7 @@ import { match } from 'ts-pattern'
 import { usePathname, useRouter } from 'next/navigation'
 import { useEffect, useState } from 'react'
 import { Plus, Filter, MegaPhone, Message, Question } from '../../assets'
-import { SelectFilterContainer } from '../../pages/lecture/style'
+import { SelectFilterContainer } from '../../../../../packages/inside/src/styles/lecture/style'
 import { LectureTypeModal } from '../../modals'
 import { useRecoilState } from 'recoil'
 import { LectureTypeText } from '../../atoms'
